renamed colorHistogramYCgCo
